requirements.txt
# Core Django
Django>=5.2,<6.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
whitenoise>=6.6,<7.0
dj-database-url>=2.1,<3.0

# Authentication
django-globus-portal-framework>=4.2,<5.0
social-auth-app-django>=5.0,<6.0

# Data pipeline (optional — needed for PubMed/GEO updates)
biopython>=1.84,<2.0
requests>=2.31,<3.0

# AI chat (optional — needed for /chat/ feature)
anthropic>=0.31,<1.0
openai>=1.0,<2.0
